requirements.txt
requests==2.32.3
Django==5.1.5
pillow==11.1.0
drf-yasg==1.21.8
django-filter==24.3
python-dotenv==1.0.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
